test that arbitrary files cannot be checksummed
New test asserts that only the checksums for retrieved files will be
checked. If a checksum file references other files, those will not be
checked.

This is an important security assertion as checksum files should not
cause go-getter to read arbitrary paths and report if their contents
match a checksum.